backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.34
psycopg[binary]>=3.2.1
celery[redis]>=5.4.0
redis>=5.0.8
python-socketio>=5.11.4
python-multipart>=0.0.9
minio>=7.2.8
pydantic-settings>=2.4.0
moviepy>=1.0.3
requests>=2.32.3
prometheus-client>=0.21.0
